added CommentsList component & fixed db models
