+++ client/src/scenes/widgets/PostWidget.jsx
@@ -6,26 +6,17 @@ import {
   EditOutlined,
   BorderColor,
   Delete,
-  SendRounded,
 } from "@mui/icons-material";
-import {
-  Box,
-  Divider,
-  IconButton,
-  Typography,
-  InputBase,
-  useTheme,
-} from "@mui/material";
+import { Box, IconButton, Typography, useTheme } from "@mui/material";
 import FlexBetween from "components/FlexBetween";
 import Friend from "components/Friend";
 import WidgetWrapper from "components/WidgetWrapper";
 import { useState } from "react";
 import { useSelector } from "react-redux";
-import { patchPostLikes, addCommentToPost, removeCommentFromPost } from "API";
+import { patchPostLikes } from "API";
 import SkeletonLoad from "components/SkeletonLoad";
 import DropdownMenu from "components/DropdownMenu";
-import UserImage from "components/UserImage";
-import { Link } from "react-router-dom";
+import CommentsList from "components/CommentsList";
 
 const PostWidget = ({
   postId,
@@ -42,9 +33,7 @@ const PostWidget = ({
   isEdited,
   isPostLoading,
 }) => {
-  const [isComments, setIsComments] = useState(false);
-  const [commentText, setCommentText] = useState("");
-  const [commentList, setCommentList] = useState(comments);
+  const [isCommentsShown, setIsCommentsShown] = useState(false);
   const token = useSelector((state) => state.auth.token);
   const loggedInUserId = useSelector((state) => state.auth.user._id);
   const [postLikes, setPostLikes] = useState(likes);
@@ -61,7 +50,6 @@ const PostWidget = ({
   const main = palette.neutral.main;
   const primary = palette.primary.main;
   const medium = palette.neutral.medium;
-  const light = palette.neutral.light;
 
   const postDropdownMenuItems = [
     {
@@ -82,21 +70,6 @@ const PostWidget = ({
   const patchLike = async () => {
     const { likes } = await patchPostLikes(postId, token, loggedInUserId);
     setPostLikes(likes);
-  };
-
-  const handleComment = async () => {
-    if (commentText.length === 0) return;
-    const formData = new FormData();
-    formData.append("userId", loggedInUserId);
-    formData.append("commentText", commentText);
-    const post = await addCommentToPost(postId, formData, token);
-    setCommentList(post.comments);
-    setCommentText("");
-  };
-
-  const handleRemoveComment = async (id) => {
-    const post = await removeCommentFromPost(postId, id, loggedInUserId, token);
-    if (post) setCommentList(post.comments);
   };
 
   return (
@@ -155,10 +128,10 @@ const PostWidget = ({
           </FlexBetween>
 
           <FlexBetween gap="0.3rem">
-            <IconButton onClick={() => setIsComments(!isComments)}>
+            <IconButton onClick={() => setIsCommentsShown(!isCommentsShown)}>
               <ChatBubbleOutlineOutlined />
             </IconButton>
-            <Typography>{commentList.length}</Typography>
+            <Typography>{comments.length}</Typography>
           </FlexBetween>
         </FlexBetween>
 
@@ -175,122 +148,11 @@ const PostWidget = ({
         </FlexBetween>
       </FlexBetween>
 
-      {isComments && (
-        <Box mt="0.5rem">
-          {commentList.map((comment) => (
-            <Box key={comment._id}>
-              <Divider sx={{ margin: "1rem 0" }} />
-
-              <Box display="flex" justifyContent="space-between">
-                <Box
-                  display="flex"
-                  justifyContent="flex-start"
-                  alignItems="flex-start"
-                >
-                  <UserImage image={comment.userPicturePath} size="45px" />
-                  <Box display="flex" flexDirection="column" ml="0.5rem">
-                    <Link
-                      to={`/profile/${comment.userId}`}
-                      style={{
-                        textDecoration: "none",
-                        display: "flex",
-                        alignItems: "center",
-                        color: "inherit",
-                      }}
-                    >
-                      <Box
-                        display="flex"
-                        flexDirection="column"
-                        justifyContent="center"
-                        backgroundColor={light}
-                        padding="0.5rem 1rem"
-                        borderRadius="35px"
-                        width="fit-content"
-                        height="45px"
-                      >
-                        <Typography fontSize="0.8rem">
-                          {comment.userName}
-                        </Typography>
-                        <Typography sx={{ fontSize: "0.65rem", color: medium }}>
-                          {new Date(comment.createdAt).toLocaleString(locale, {
-                            timezone,
-                            year: "numeric",
-                            month: "numeric",
-                            day: "numeric",
-                            hour: "2-digit",
-                            minute: "2-digit",
-                          })}
-                        </Typography>
-                      </Box>
-                    </Link>
-                    <Typography
-                      sx={{
-                        color: main,
-                        m: "0.5rem 0",
-                        p: "0.25rem 0.25rem 0 ",
-                        whiteSpace: "pre-line",
-                      }}
-                    >
-                      {comment.commentText}
-                    </Typography>
-                  </Box>
-                </Box>
-                {comment.userId === loggedInUserId && (
-                  <Box>
-                    <IconButton
-                      onClick={() => handleRemoveComment(comment._id)}
-                    >
-                      <Delete
-                        sx={{
-                          color: medium,
-                        }}
-                      />
-                    </IconButton>
-                  </Box>
-                )}
-              </Box>
-            </Box>
-          ))}
-          <Divider sx={{ margin: "0.75rem 0" }} />
-          <Box
-            display="flex"
-            padding="0.5rem 0 0.5rem 0.75rem"
-            borderRadius="0.5rem"
-            backgroundColor={palette.neutral.light}
-          >
-            <InputBase
-              placeholder="What's on your mind..."
-              inputProps={{
-                className: "add-post-input",
-              }}
-              onChange={(e) => setCommentText(e.target.value)}
-              value={commentText}
-              multiline={true}
-              maxRows={3}
-              sx={{
-                flex: "9",
-                paddingRight: "0.3125rem",
-              }}
-            />
-            <Box
-              sx={{
-                flex: "1",
-                display: "flex",
-                alignItems: "center",
-              }}
-            >
-              <Box>
-                <IconButton
-                  onClick={handleComment}
-                  style={{ backgroundColor: "transparent" }}
-                >
-                  <SendRounded />
-                </IconButton>
-              </Box>
-            </Box>
-          </Box>
-        </Box>
-      )}
+      <CommentsList
+        postId={postId}
+        comments={comments}
+        isShown={isCommentsShown}
+      />
     </WidgetWrapper>
   );
 };
