+++ server/models/Comment.js
@@ -3,14 +3,14 @@ import mongoose from "mongoose";
 
 const commentSchema = mongoose.Schema(
   {
-    user: {
-      type: Schema.Types.ObjectId,
-      ref: "User",
-      required: true,
-    },
     post: {
       type: Schema.Types.ObjectId,
       ref: "Post",
+      required: true,
+    },
+    author: {
+      type: Schema.Types.ObjectId,
+      ref: "User",
       required: true,
     },
     content: {
