+++ server/models/Post.js
@@ -8,17 +8,13 @@ const postSchema = mongoose.Schema(
       ref: "User",
       required: true,
     },
-
     description: String,
     picturePath: String,
     likes: {
       type: Map,
       of: Boolean,
     },
-    comments: {
-      type: Array,
-      default: [{ type: Schema.Types.ObjectId, ref: "Comment" }],
-    },
+    comments: [{ type: Schema.Types.ObjectId, ref: "Comment" }],
     isEdited: {
       type: Boolean,
       default: false,
